+++ dongmanee/src/main/java/com/dongmanee/domain/security/service/AuthService.java
@@ -1,0 +1,7 @@
+package com.dongmanee.domain.security.service;
+
+import com.dongmanee.domain.security.dto.response.JwsToken;
+
+public interface AuthService {
+	JwsToken login(String loginId, String password);
+}
