--- dongmanee/src/main/java/com/dongmanee/global/security/login/service/SignInService.java
@@ -1,7 +1,0 @@
-package com.dongmanee.global.security.login.service;
-
-import com.dongmanee.global.security.login.dto.response.JwsToken;
-
-public interface SignInService {
-    JwsToken login(String loginId, String password);
-}
